% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateIS.R
\name{annotateIS}
\alias{annotateIS}
\title{Annotate Integration Sites}
\usage{
annotateIS(IS.path = NULL, gtf.path = NULL, sample = NULL,
  mode = NULL)
}
\arguments{
\item{sample}{charachter. Name of the sample currently used.}

\item{IS}{data_frame. Integration table generated by \code{"getISposition"}}

\item{geneBedFile.path}{charachter. Path to the sorted annotation file.}
}
\value{
a tibble, annotated.
}
\description{
Given a sorted annotation file generated with 'clonality_indexBuilder.sh' (chr, start, end, strand, ensembl_id, gene_name), annotate integration sites.
Returns information about the closest features, distances, type of IS (genic or intergenic) and context (concordant or discordant to the nearest up/down features)
}
\examples{
LTR3.positions <- getISposition(r1 = r1.LTR3.tidy, r2 = r2.LTR3.tidy, randomTag = randomTag, LTR = "LTR3", recall = FALSE, maxgap = 75)
annotateIS(IS = LTR3.positions, geneBedFile = "path/to/geneInfos.sorted.txt", sample = "mySample")

}
\author{
Vincent Hahaut & Nicolas Rosewick
}
