% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IStoGTF.R
\name{IStoGTF}
\alias{IStoGTF}
\title{Transform an IS table into GTF}
\usage{
IStoGTF(IS = NULL, win = 1, name = "mySample")
}
\arguments{
\item{IS}{tibble. Integration site tibble.}

\item{win}{numeric. up/down window to add to the position in order to see them in IGV}

\item{mode}{character. "IS" or "HS" for integration sites or hotspots}
}
\value{

}
\description{
Transform an IS table into GTF
}
\examples{
VIDO_createGTF_IS()

}
\author{
Vincent Hahaut
}
