% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseGTF.R
\name{parseGTF}
\alias{parseGTF}
\title{Read a GTF file into R}
\usage{
parseGTF(gtf = NULL, type = "exon", prefix = "OAR")
}
\arguments{
\item{gtf}{character. Path to the GTF file.}

\item{type}{character. Which feature should you extract}

\item{prefix}{character. Prefix to add to the chromosome names}
}
\value{

}
\description{
Read a GTF file into R
}
\examples{
parseGTF()

}
\author{
Vincent Hahaut & Nicolas Rosewick
}
