% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCIP_filter.R
\name{PCIP_filter}
\alias{PCIP_filter}
\title{Filter Pairwise Alignment File produced by Minimap2}
\usage{
PCIP_filter(minimap2PAF = NULL, targetName = NULL)
}
\arguments{
\item{minimap2PAF}{tibble. Pairwise Alignment file generated by minimap2 and loaded with \code{readPairwiseAlignmentFile}}

\item{targetName}{character. Name of the TARGET chromosome}
}
\value{
PCIP filtered and formated PAF tibble. Field description:
\itemize{
\item seqnames (character) Chromosome name
\item start (integer) Substring mapping start position
\item end (integer) Substring mapping end position
\item readStart_position (integer) Start position of the substring within the read
\item readEnd_position (integer) End position of the substring within the read
\item width (integer) Width of the substring
\item strand (character) Mapping strand (+ or -)
\item ligation (character) TARGET-HOST structure of the read as binary code (0 = HOST and 1 = TARGET)
\item mapGap (integer) Length of the gap between the end of this mapped substring and the next mapped read substring
\item readID (character) ID of the read
\item qLength (integer) Total length of the read
\item target (logical) TARGET chromosome
\item meanGap (integer) Mean of the read mapGap
\item sdGap (integer) Standard deviation of the read mapGap
\item readSide (character) Describe the position of the substring related to the entire read. Left as first 10% or else right. Used to catch the exact breakpoint.
}
}
\description{
Filter out secundary alignments and non-TARGET/HOST chimeric.
In PAF format each line represents the mapping information of a read substring.
While filtering an extra field called "ligation" is created depicting the presence/absence of mapping onto the TARGET using the following code: 0 (mapping HOST) and 1 (mapping TARGET). This field is collapsed by read and only reads harboring one of the following substrings combinations are selected: 01, 10, 101, 1001.
}
\note{
minimap2PAF <- readPairwiseAlignmentFile(minimap2PAF="my/path/to/minimap2.align")
PCIP_filter(minimap2PAF = minimap2PAF, targetName="HTLV")
}
\author{
Vincent Hahaut
}
\keyword{PCIP}
