% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTargetHostSequence.R
\name{PCIP_getTargetHostSequence}
\alias{PCIP_getTargetHostSequence}
\title{PCIP_getTargetHostSequence}
\usage{
PCIP_getTargetHostSequence(fasta.genome = NULL, fasta.target = NULL,
  position = NULL, size = 10000, orientation = "+",
  out.path = "~/Desktop/")
}
\arguments{
\item{fasta.genome}{Character. Path/to/genome.fa.}

\item{fasta.target}{Character. Path/to/target.fa.}

\item{position}{GRanges compatible position. "chrZ:start-end", position where the target is inserted.}

\item{size}{Numeric. Size of the genomic window to consider.}

\item{orientation}{"+" or "-". Reverse the target sequence if "-".}

\item{out.path}{Character. Path/to/, where the new fasta reference will be written.}
}
\value{
Write a FASTA file containing the proviral sequence inserted in the middle of the genomic sequence of interest.
}
\description{
Return the target sequence surrounded by the genomic sequence into which it is inserted
}
\examples{
PCIP_getTargetHostSequence()

}
\author{
Vincent Hahaut
}
