% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCIP_targetSV.R
\name{PCIP_targetSV}
\alias{PCIP_targetSV}
\title{Call deletions from PCIP-seq Nanopore data}
\usage{
PCIP_targetSV(bam.path = NULL, viralLength = NULL, minseglen = 50,
  out = NULL, minCoverage = 5, ID = NULL, provirus = "all",
  results = "FILTER")
}
\arguments{
\item{bam.path}{character. Path/to/myfile.sorted.bam. Index (.bai, Samtools) is expected in the same folder}

\item{viralLength}{numeric. Size of the provirus}

\item{minseglen}{numeric. Minimal size of the deletion to be found (shorter than 10 is not recommended)}

\item{out}{character. Path/to/output/prefix}

\item{minCoverage}{numeric. Minimum coverage to call deletion}

\item{ID}{character. Sample ID}

\item{provirus}{character. Can be set to the TARGET chromosome name to call deletion only in the TARGET or "all" to avoid any pre-filtering}

\item{results}{character. If "FILTER", only results passing the FILTER threshold will be returned in an aggregated format.}
}
\value{
graphic. Depicts Base-Coverage and Deletion coverage accros the sequence. Significant change points are marked in red and deletions are highlighted in black.

tibble. Deletion table report.
\itemize{
\item CHROM: character. Chromosome name
\item START: integer. Window start position
\item END: integer. Window end position
\item WIDTH: integer. Window width
\item BASE.SUM: integer. Number of sequenced "ATCG"
\item BASE.COVERAGE: double. "ATCG" coverage
\item DEL.SUM: integer. Number of sequenced "-"
\item DEL.COVERAGE: double. "-" coverage
\item COVERAGE: double. Total coverage of the window ("ATCG", "+", "-")
\item CATEGORY: character. Currently only "DEL" category exists
\item RATIO: double. Ratio between DEL.COVERAGE and BASE.COVERAGE.
\item FILTER: character. COVERAGE (coverage is lower than minCoverage), SHORT (the segment is smaller than minseglen), OUT (Base / deletion ratio is bigger than 5), PASS (Base / deletion ratio is smaller than .8), AMBIGUIOUS (other cases to be manually checked).
\item ID: character. Sample ID
}
}
\description{
Primary function of this tool is to detect deletions lying within defined genomic segment.
Sudden drops of mean and variance of coverage is monitored through a change point PELT mean-var function.
Decreases and increases are use to delimit windows.
Windows harboring a ratio deletion / coverage of more than 80% are flagged as DELETION.
}
\author{
Vincent Hahaut
}
\keyword{PCIP}
